requirements.txt
aiohttp==1.0.5
async-timeout==1.1.0
chardet==2.3.0
discord.py==0.13.0
multidict==2.1.2
requests==2.11.1
websockets==3.2
